Add a --force flag to kill_all_nodes

+++ roles/slurm/files/kill_all_nodes.sh
@@ -3,13 +3,15 @@ set -euo pipefail
 
 # TODO Use python-citc
 
-read -p "Are you sure you want to kill all running compute nodes [y/N]? " -r
-if [[ ! $REPLY =~ ^[Yy]$ ]]
-then
-    echo Exiting without killing nodes
-    exit 1
+if [[ $# -ne 1 || "${1}" != "--force" ]]; then
+    read -p "Are you sure you want to kill all running compute nodes [y/N]? " -r
+    if [[ ! $REPLY =~ ^[Yy]$ ]]
+    then
+        echo Exiting without killing nodes
+        exit 1
+    fi
+    echo
 fi
-echo
 
 echo Terminating any remaining compute nodes
 if systemctl status slurmctld >> /dev/null; then
